apps/python-demo/pyproject.toml
[project]
name = "mvp_agent_demo"
version = "0.2.0"
description = "MVP v0.2 demo: Plan → Execute → Gate (telemetry-driven)"
requires-python = ">=3.10"
dependencies = [
  "traceloop-sdk>=0.26.0",
  "opentelemetry-api>=1.24.0",
]

[project.scripts]
mvp-agent-demo = "mvp_agent.main:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
